added symmetrization. works like a charm now!

+++ FMCA/src/util/Macros.h
@@ -21,7 +21,7 @@ namespace FMCA {
 #define FMCA_BBOX_THREASHOLD 5e-2
 #define FMCA_ZERO_TOLERANCE 1e-14
 
-typedef unsigned int IndexType;
+typedef unsigned long long int IndexType;
 } // namespace FMCA
 
 #endif
